M8T1 memo & useCallback
