+++ src/components/main/map.tsx
@@ -51,7 +51,7 @@ function Map(props: MapProps): JSX.Element {
           .addTo(markerLayer);
       });
 
-      //map.flyTo({lat: city.location.latitude, lng: city.location.longitude}, city.location.zoom);
+      //TODO map.flyTo({lat: city.location.latitude, lng: city.location.longitude}, city.location.zoom);
 
       return () => {
         map.removeLayer(markerLayer);
